refact: auth 폴더 삭제후 유저로 이동
