+++ src/user/user.module.ts
@@ -4,14 +4,26 @@ import { UserRepository } from './user.repository';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { TypeOrmExModule } from '../common/decorator/typeorm-ex.module'
-import { User } from './entities/user.entity';
+import { PassportModule } from '@nestjs/passport';
+import * as config from 'config'
+import { JwtModule } from '@nestjs/jwt';
+import { JwtStrategy } from './jwt.strategy';
+
+const jwtConfig = config.get('jwt')
 
 @Module({
-    imports:[
-        TypeOrmModule.forFeature([User]),
+    imports: [
+        PassportModule.register({ defaultStrategy: 'jwt' }),
+        JwtModule.register({
+          secret: process.env.JWT_SECRET || jwtConfig.secret,
+          signOptions: {
+            expiresIn: jwtConfig.expiresIn,
+          }
+        }),
         TypeOrmExModule.forCustomRepository([UserRepository])
-    ],
+      ],
     controllers:[UserController],
-    providers:[UserService]
+    providers: [UserService, JwtStrategy],
+  exports: [JwtStrategy, PassportModule]
 })
 export class UserModule {}
